+++ inbm-vision/Changelog.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## NEXT - ?
+
+### Fixed
+- 504494, HSD 15010332252 - Fix provision file not removed in /opt/xlink_provision when failed to reset device due to no matching device id
+- 506035, HSD 15010309026 - Update node apparmor policy by adding /sys/kernel/xlink_secure/debug
+
 ## 3.0.6 - 2021-12-31
 This is the first open source release.